+++ lib/pages/auth/regist_page.dart
@@ -10,7 +10,7 @@ import 'package:toefl/utils/hex_color.dart';
 import 'package:toefl/widgets/blue_button.dart';
 import 'package:toefl/widgets/form_input.dart';
 
-import '../remote/api/user_api.dart';
+import '../../remote/api/user_api.dart';
 
 class RegistPage extends StatefulWidget {
   const RegistPage({super.key});
