+++ lib/routes/routes.dart
@@ -5,7 +5,7 @@ import 'package:toefl/models/quiz_type.dart';
 import 'package:toefl/pages/games/quiz/finish_quiz_page.dart';
 import 'package:toefl/pages/games/quiz/init_quiz.dart';
 import 'package:path/path.dart';
-import 'package:toefl/pages/forgot_password_page.dart';
+import 'package:toefl/pages/auth/forgot_password_page.dart';
 import 'package:toefl/pages/games/quiz/grammar_page.dart';
 import 'package:toefl/pages/games/quiz/quiz_page.dart';
 import 'package:toefl/pages/bookmark/bookmark_detail_page.dart';
@@ -19,21 +19,21 @@ import 'package:toefl/pages/mini_test/mini_test_page.dart';
 import 'package:toefl/pages/games/quiz/quiz_page.dart';
 import 'package:toefl/pages/user/edit_profile_page.dart';
 import 'package:toefl/pages/full_test/opening_loading_page.dart';
-import 'package:toefl/pages/new_password_page.dart';
-import 'package:toefl/pages/otp_verification_page.dart';
+import 'package:toefl/pages/auth/new_password_page.dart';
+import 'package:toefl/pages/auth/otp_verification_page.dart';
 import 'package:toefl/pages/user/profile_page.dart';
 import 'package:toefl/pages/full_test/full_test_page.dart';
 import 'package:toefl/pages/full_test/simulation_page.dart';
 import 'package:toefl/pages/full_test/test_result_page.dart';
-import 'package:toefl/pages/login_page.dart';
+import 'package:toefl/pages/auth/login_page.dart';
 import 'package:toefl/pages/on_boarding.dart';
 import 'package:toefl/pages/rank_page.dart';
-import 'package:toefl/pages/regist_page.dart';
+import 'package:toefl/pages/auth/regist_page.dart';
 import 'package:toefl/pages/games/games_page.dart';
 import 'package:toefl/pages/review_test/review_test_page.dart';
-import 'package:toefl/pages/setgoal_page.dart';
+import 'package:toefl/pages/user/setgoal_page.dart';
 import 'package:toefl/pages/splash_page.dart';
-import 'package:toefl/pages/success_password_page.dart';
+import 'package:toefl/pages/auth/success_password_page.dart';
 import 'package:toefl/pages/template_page.dart';
 import 'package:toefl/routes/route_key.dart';
 import 'package:toefl/pages/main_page.dart';
